refactor: 공통 컴포넌트로 이동
