+++ components/ViewAnswer.tsx
